pyproject.toml
[tool.poetry]
name = "transit-chem"
version = "0.6.1"
description = "Quantifying Probabilistic Electron Transit times."
authors = ["Evan Curtin <fakeemail@gmail.com>"]
readme = "README.rst"
license = "MIT"

[tool.poetry.scripts]
transit-chem = 'transit_chem.cli:cli'

[tool.poetry.dependencies]
python = "^3.7.0"
click = "^7.0"
numba = "^0.43.0"
numpy = "^1.17"
scipy = "^1.3"
attrs = "^19.3"
tqdm = "^4.39"
loguru = "^0.4.0"
joblib = "^0.14.1"
matplotlib = "^3.1"
cattrs = "^1.0"


[tool.poetry.dev-dependencies]
pytest = "^3.4"
black = { version = "*", allows-prereleases = true }
pytest-cov = "^2.8"
isort = "^4.3"
hypothesis = "^4.42"
flake8 = "^3.7"
pip = "^19.3"
coverage = "^4.5"
mypy = "^0.740.0"
IPython = "^7.9"
nbsphinx = "^0.4.3"
sphinx = "^2.2"
sphinx-autodoc-typehints = "^1.9"
sphinx_bootstrap_theme = "^0.7.1"
jupyter = "^1.0"
jupyterlab = "^1.2"
py-spy = "^0.3.1"


[tool.black]
line-length = 100
target-version = ['py37']